Read işlemleri tamamlandı.
Verileri okuma-detayları görüntüleme işlemleri tamamlandı.
Scope türleri ViewScoped olarak değiştirildi.

+++ src/main/java/org/freesoftware/dao/PersonDAO.java
@@ -13,6 +13,6 @@ public interface PersonDAO {
 
 	public Person updatePerson(Person person);
 
-	public Person getPerson(int personId);
+	public Person getPerson(long personId);
 
 }
